requirements.txt
notebook
torchvision
mplfinance
yfinance
tensorboard
pyts
ruptures
